Transcriber/src/requirements.txt
faster-whisper>=1.0.0
openai-whisper>=20240930
pywhispercpp>=1.2.0
watchdog>=4.0.0
psutil>=5.9.0
PyYAML>=6.0
